# openapi.json
{
...
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/v1/sessions": {
      "post": {
        "operationId": "createSession",
        "summary": "",
        "description": "Create session",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The session has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        },
        "tags": ["session"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "listSessions",
        "summary": "",
        "description": "List sessions",
        "parameters": [
          {
            "name": "_sort",
            "required": false,
            "in": "query",
            "description": "排序参数",
            "schema": {
              "enum": ["createAt", "-createAt", "updateAt", "-updateAt", "expireAt", "-expireAt"],
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "description": "用 key 进行查询",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "client",
            "required": false,
            "in": "query",
            "description": "客户端/设备",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tokenExpiresIn",
            "required": false,
            "in": "query",
            "description": "token 有效时长\n\nshort time span string\n\nrefs: https://github.com/vercel/ms\n\neg: \"2 days\", \"10h\", \"7d\", \"120s\", \"2.5 hrs\", \"2h\", \"1m\", \"5s\", \"1y\", \"100\", \"1y1m1d\"\n\nm => minute\nh => hour\nd => day\nw => week\nM => month\ny => year\ns => second\nms => millisecond\n无单位 => millisecond",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": false,
            "in": "query",
            "description": "用户 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_limit",
            "required": false,
            "in": "query",
            "description": "分页大小",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "_offset",
            "required": false,
            "in": "query",
            "description": "分页偏移",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paged array of sessions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  }
                }
              }
            }
          }
        },
        "tags": ["session"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/v1/sessions/{sessionId}": {
      "get": {
        "operationId": "getSession",
        "summary": "",
        "description": "Find session by id",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The session with expected id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        },
        "tags": ["session"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "updateSession",
        "summary": "",
        "description": "Update session",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The session updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        },
        "tags": ["session"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteSession",
        "summary": "",
        "description": "Delete session",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content."
          }
        },
        "tags": ["session"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/v1/sessions/@restrictToken": {
      "post": {
        "operationId": "restrictToken",
        "summary": "",
        "description": "返回一个 token\n\n一般用于给设备等提供临时访问凭证",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestrictTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlyToken"
                }
              }
            }
          },
          "201": {
            "description": "The restricted token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlyToken"
                }
              }
            }
          }
        },
        "tags": ["session"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Auth API Server",
    "description": "Auth API for auth service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "IdentityType": {
        "type": "string",
        "description": "认证类型",
        "enum": ["ID"]
      },
      "Identity": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/IdentityType"
          },
          "name": {
            "type": "string",
            "description": "真实姓名"
          },
          "verifyAt": {
            "format": "date-time",
            "type": "string",
            "description": "认证时间"
          },
          "verified": {
            "type": "boolean",
            "description": "是否认证通过"
          }
        },
        "required": ["verified"]
      },
      "User": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "密码",
            "writeOnly": true
          },
          "avatar": {
            "type": "string",
            "description": "头像"
          },
          "intro": {
            "type": "string",
            "description": "简介"
          },
          "data": {
            "type": "string",
            "description": "额外数据"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "language": {
            "type": "string",
            "description": "使用语言"
          },
          "lastSeenAt": {
            "format": "date-time",
            "type": "string",
            "description": "最后活跃时间"
          },
          "lastLoginIp": {
            "type": "string",
            "description": "最后登录 IP"
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "ns": {
            "type": "string",
            "description": "所属命名空间"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "registerIp": {
            "type": "string",
            "description": "注册 IP"
          },
          "registerRegion": {
            "type": "string",
            "description": "注册地区，存地区编号"
          },
          "roles": {
            "description": "角色",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "super": {
            "type": "boolean",
            "description": "是否超级管理员"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "dialingPrefix": {
            "type": "string",
            "description": "区号"
          },
          "identity": {
            "$ref": "#/components/schemas/Identity"
          },
          "id": {
            "type": "string",
            "description": "Entity id"
          },
          "createAt": {
            "format": "date-time",
            "type": "string",
            "description": "Entity created at when"
          },
          "updateAt": {
            "format": "date-time",
            "type": "string",
            "description": "Entity updated at when"
          },
          "createBy": {
            "type": "string",
            "description": "Entity created by who"
          },
          "updateBy": {
            "type": "string",
            "description": "Entity updated by who"
          }
        },
        "required": ["ns", "id"]
      },
...
      "CreateSessionDto": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "description": "用户 ID"
          },
          "acl": {
            "type": "object",
            "description": "访问控制列表"
          },
          "expireAt": {
            "format": "date-time",
            "type": "string",
            "description": "会话过期时间"
          },
          "client": {
            "type": "string",
            "description": "客户端/设备"
          },
          "tokenExpiresIn": {
            "type": "string",
            "description": "token 有效时长\n\nshort time span string\n\nrefs: https://github.com/vercel/ms\n\neg: \"2 days\", \"10h\", \"7d\", \"120s\", \"2.5 hrs\", \"2h\", \"1m\", \"5s\", \"1y\", \"100\", \"1y1m1d\"\n\nm => minute\nh => hour\nd => day\nw => week\nM => month\ny => year\ns => second\nms => millisecond\n无单位 => millisecond"
          }
        },
        "required": ["uid", "expireAt"]
      },
      "Session": {
        "type": "object",
        "properties": {
          "acl": {
            "type": "object",
            "description": "访问控制列表"
          },
          "expireAt": {
            "format": "date-time",
            "type": "string",
            "description": "会话过期时间"
          },
          "key": {
            "type": "string",
            "description": "refresh token key"
          },
          "user": {
            "description": "用户 ID",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "client": {
            "type": "string",
            "description": "客户端/设备"
          },
          "tokenExpiresIn": {
            "type": "string",
            "description": "token 有效时长\n\nshort time span string\n\nrefs: https://github.com/vercel/ms\n\neg: \"2 days\", \"10h\", \"7d\", \"120s\", \"2.5 hrs\", \"2h\", \"1m\", \"5s\", \"1y\", \"100\", \"1y1m1d\"\n\nm => minute\nh => hour\nd => day\nw => week\nM => month\ny => year\ns => second\nms => millisecond\n无单位 => millisecond"
          },
          "id": {
            "type": "string",
            "description": "Entity id"
          },
          "createAt": {
            "format": "date-time",
            "type": "string",
            "description": "Entity created at when"
          },
          "updateAt": {
            "format": "date-time",
            "type": "string",
            "description": "Entity updated at when"
          },
          "createBy": {
            "type": "string",
            "description": "Entity created by who"
          },
          "updateBy": {
            "type": "string",
            "description": "Entity updated by who"
          }
        },
        "required": ["expireAt", "key", "user", "id"]
      },
      "UpdateSessionDto": {
        "type": "object",
        "properties": {
          "acl": {
            "type": "object",
            "description": "访问控制列表"
          },
          "expireAt": {
            "format": "date-time",
            "type": "string",
            "description": "会话过期时间"
          },
          "client": {
            "type": "string",
            "description": "客户端/设备"
          },
          "tokenExpiresIn": {
            "type": "string",
            "description": "token 有效时长\n\nshort time span string\n\nrefs: https://github.com/vercel/ms\n\neg: \"2 days\", \"10h\", \"7d\", \"120s\", \"2.5 hrs\", \"2h\", \"1m\", \"5s\", \"1y\", \"100\", \"1y1m1d\"\n\nm => minute\nh => hour\nd => day\nw => week\nM => month\ny => year\ns => second\nms => millisecond\n无单位 => millisecond"
          },
          "uid": {
            "type": "string",
            "description": "用户 ID"
          }
        }
      },
      "RestrictTokenDto": {
        "type": "object",
        "properties": {
          "expiresIn": {
            "type": "string",
            "description": "short time span string\n\nrefs: https://github.com/vercel/ms\n\neg: \"2 days\", \"10h\", \"7d\", \"120s\", \"2.5 hrs\", \"2h\", \"1m\", \"5s\", \"1y\", \"100\", \"1y1m1d\"\n\nm => minute\nh => hour\nd => day\nw => week\nM => month\ny => year\ns => second\nms => millisecond\n无单位 => millisecond"
          },
          "acl": {
            "type": "object",
            "description": "访问控制列表"
          },
          "key": {
            "type": "string",
            "description": "refresh token key"
          },
          "ns": {
            "type": "string",
            "description": "命名空间约束，只能是 token 的同级或者下一级"
          }
        },
        "required": ["expiresIn", "key"]
      },
      "OnlyToken": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "token"
          },
          "tokenExpireAt": {
            "format": "date-time",
            "type": "string",
            "description": "token 过期时间"
          }
        },
        "required": ["token", "tokenExpireAt"]
      },
...
    }
  }
}